Closes #3030 numpy alignment directory structure
