+++ PROTO_tests/tests/numpy/numpy_test.py
@@ -1,0 +1,2 @@
+class TestStats:
+    pass
