+++ arkouda/scipy/__init__.py
