+++ arkouda/scipy/special/__init__.py
