+++ pytest.ini
@@ -2,8 +2,9 @@
 filterwarnings =
     ignore:Version mismatch between client .*
 testpaths =
-    tests/akscipy/math_test.py
-    tests/akscipy/akscipy_test.py
+    tests/scipy/scipy_test.py
+    tests/scipy/special_test.py
+    tests/numpy/numpy_test.py
     tests/alignment_tests.py
     tests/array_view_test.py
     tests/bitops_test.py
