+++ pytest_PROTO.ini
@@ -7,6 +7,8 @@ filterwarnings =
     ignore:Version mismatch between client .*
 testpaths =
     tests/*_test
+    tests/scipy/test
+    tests/numpy/test
 ;    tests/benchmarks/*_benchmark.py
 norecursedirs =
     .git
