+++ tests/numpy/numpy_test.py
@@ -1,0 +1,6 @@
+from base_test import ArkoudaTest
+
+
+class StatsTest(ArkoudaTest):
+    def setUp(self):
+        ArkoudaTest.setUp(self)
